+++ src/chunkinator/image_sampling.h
@@ -7,4 +7,4 @@
 
 using namespace godot;
 
-double bilinearly_sample_image_single_channel(Ref<Image> p_image, int p_channel_idx, const Vector2 &p_uv);
+float bilinearly_sample_image_single_channel(Ref<Image> p_image, int p_channel_idx, const Vector2 &p_uv);
